% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{addisoinfo}
\alias{addisoinfo}
\title{Add ISO information}
\usage{
addisoinfo(datatable, datecolumn = "DATE", timezone = "")
}
\arguments{
\item{datatable}{The data.table to add ISO weeknumber, year, and month to.}

\item{datecolumn}{String identifying the column name in which a date is present. The column itself should be
of the POSIXct class. Default=\code{"DATE"}.}

\item{timezone}{String determining the timezone (from \code{\link[base]{strptime}}). Default=\code{""},
which uses the local time zone.}
}
\value{
Returns the modified data.table invisibly (modified by reference)
with the added columns ISOWEEKNR, ISOYEAR and ISOMONTH.
}
\description{
Add ISO date information to a data.table
}
\seealso{
Other data.table functions: 
\code{\link{addcolumn}()},
\code{\link{copytoclipboard}()},
\code{\link{dojoin}()},
\code{\link{extractcategories}()},
\code{\link{fulljoin}()},
\code{\link{getcolumnclasses}()},
\code{\link{innerjoin}()},
\code{\link{leftjoin}()},
\code{\link{loadlargetable}()},
\code{\link{loadsastable}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{makeintegercolumns}()},
\code{\link{makenumericcolumns}()},
\code{\link{mapcolumnclasses}()},
\code{\link{prefixcolnames}()},
\code{\link{removeNAfromtable}()},
\code{\link{removeallNAcolumns}()},
\code{\link{removeallNArows}()},
\code{\link{removecolumns}()},
\code{\link{removeisoinfo}()},
\code{\link{removerows}()},
\code{\link{renamecolumns}()},
\code{\link{rightjoin}()},
\code{\link{runcolfun}()},
\code{\link{selectuniques}()},
\code{\link{selectwherein}()},
\code{\link{selectwhere}()},
\code{\link{setcolclass}()},
\code{\link{setcolnames}()},
\code{\link{setcoltofront}()},
\code{\link{setcolumns}()},
\code{\link{unlistcolumns}()},
\code{\link{updatetablevalues}()}

Other date and time functions: 
\code{\link{makeconsistentdateformat}()},
\code{\link{makeconsistentdatetimeformat}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{removeisoinfo}()},
\code{\link{seqdays}()}
}
\concept{data.table functions}
\concept{date and time functions}
