% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedConversion.R
\name{bitsinbinary}
\alias{bitsinbinary}
\title{Find bits in binary}
\usage{
bitsinbinary(binaryvalues)
}
\arguments{
\item{binaryvalues}{String vector containing binary bits.}
}
\value{
List with the locations which bits are set to 1. If no bits are set,
or the provided binary value is invalid, the list value contains \code{integer(0)}.
}
\description{
Find which bits in a binary string are set to 1.
}
\examples{
bitsinbinary(c("0010","0110"))
bitsinbinary(c("1010","string","0110"))
}
\seealso{
Other conversion functions: 
\code{\link{convertcoords_degrees2decimal}()},
\code{\link{hextobinary}()}
}
\concept{conversion functions}
