% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedVector.R
\name{chunk}
\alias{chunk}
\title{Create chunks}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{The vector or list to split}

\item{n}{The maximum length the vector or list is allowed to be.}
}
\value{
Vector or list split into chunks, as unnamed lists.
}
\description{
Split a vector or list into chunks of a provided length.
}
\seealso{
Other vector functions: 
\code{\link{dorollapply}()},
\code{\link{extendvector}()},
\code{\link{findconstantchunks}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{replaceelements}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()},
\code{\link{shift}()}
}
\concept{vector functions}
