% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedConversion.R
\name{convertcoords_degrees2decimal}
\alias{convertcoords_degrees2decimal}
\title{Converts degrees-minutes}
\usage{
convertcoords_degrees2decimal(degrees)
}
\arguments{
\item{degrees}{The input longitude or latitude in degrees-minutes.}
}
\value{
The longitude or latitude in degrees-decimals.
}
\description{
Converts longitude or latitude in degrees-minutes to degrees-decimals.
}
\seealso{
Other conversion functions: 
\code{\link{bitsinbinary}()},
\code{\link{hextobinary}()}
}
\concept{conversion functions}
