% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedData.R
\name{createbins}
\alias{createbins}
\title{Create time series bins}
\usage{
createbins(thetimes, thedata = NULL, binsize = "min")
}
\arguments{
\item{thetimes}{Data.table with the first column being the DateTime and the second column being the values,
or, a vector containing the DateTime only.}

\item{thedata}{Vector containing the values correlated to the DateTime. Must be of equal length to \code{thetimes}.
Default=\code{NULL}, which only uses the first parameter.}

\item{binsize}{The bin size to split by. Can be \code{sec}, \code{min}, \code{hour}, \code{mday}, \code{mon}, \code{year},
\code{wday} or \code{yday}.}
}
\value{
Data.table with the starting times of the bins and each column with statistical information about that bin.
}
\description{
Create time series bin by calculating statistical information for each bin containing data. It will use the rounded times
as start and end times, even if data is starting at the end of the bin. There must be both time and data provided.
}
\seealso{
Other data functions: 
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{removeoffset}()},
\code{\link{timeseriesstats}()}

Other time series functions: 
\code{\link{removeoffset}()},
\code{\link{timeseriesstats}()}
}
\concept{data functions}
\concept{time series functions}
