% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedMath.R
\name{deg2rad}
\alias{deg2rad}
\title{Degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Degrees.}
}
\value{
Radians.
}
\description{
Degrees to radians
}
\seealso{
Other math functions: 
\code{\link{rad2deg}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()}
}
\concept{math functions}
