% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedStrings.R
\name{domain}
\alias{domain}
\title{Get domain
Get domain without http or www elements}
\usage{
domain(x)
}
\arguments{
\item{x}{Vector containing the strings to get domains from}
}
\value{
Vector of the domains
}
\description{
Get domain
Get domain without http or www elements
}
\seealso{
Other string functions: 
\code{\link{paste0collapse}()},
\code{\link{str_pad_month}()}
}
\concept{string functions}
