% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedVector.R
\name{dorollapply}
\alias{dorollapply}
\title{Rolling apply}
\usage{
dorollapply(vec, width, FUN)
}
\arguments{
\item{vec}{The vector to apply the function on.}

\item{width}{Width (or length) of the window.}

\item{FUN}{The function to roll over the vector. Calls the function for
all length pieces.}
}
\description{
Apply a function rolling over vector data.
}
\seealso{
Other vector functions: 
\code{\link{extendvector}()},
\code{\link{findconstantchunks}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{replaceelements}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()},
\code{\link{shift}()}
}
\concept{vector functions}
