% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedVector.R
\name{extendvector}
\alias{extendvector}
\title{Extend vector}
\usage{
extendvector(vec, len, extval = NA)
}
\arguments{
\item{vec}{The vector to set the length for.}

\item{len}{The length the vector needs to be.}

\item{extval}{The value to use in case the provided \code{vec} is shorter
than the required \code{len}. Default=\code{NA}.}
}
\value{
The vector having the required lenght, extended if needed.
}
\description{
Ensure a certain length of the vector is maintained, by cutting it
short or extending it to the required length with a specific single value.
}
\seealso{
Other vector functions: 
\code{\link{dorollapply}()},
\code{\link{findconstantchunks}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{replaceelements}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()},
\code{\link{shift}()}
}
\concept{vector functions}
