% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedVector.R
\name{findconstant}
\alias{findconstant}
\title{Find constant value}
\usage{
findconstant(
  thevector,
  minpoints = 20,
  stdlevel = 0.05,
  desiredvalue = NULL,
  desiredvaluecloseness = 0.05
)
}
\arguments{
\item{thevector}{Numeric vector to search in.}

\item{minpoints}{Length of the window that requires to be constant. Default=\code{20}.}

\item{stdlevel}{The maximum standard deviation of the window. Any standard
deviation above this level will be considered not constant. Default=\code{0.05}.}

\item{desiredvalue}{If provided, it will search for the vector windows for the closest
desired average value with a standard deviation below the provided \code{stdlevel}.
Default=\code{NULL}, which omits the search for a desired value.}

\item{desiredvaluecloseness}{Value to determine if the provided \code{desiredvalue}
is close enough to the value found. It will check if the window average is between
\code{desiredvalue-desiredvaluecloseness} and \code{desiredvalue+desiredvaluecloseness} to
be a match. Default=\code{0.05}.}
}
\value{
A list with the following items:
\describe{
\item{locations}{If \code{desiredvalue} is provided, the vector locations which result in the
closest average \code{desiredvalue} with at least the standard deviation below \code{stdlevel}.
Otherwise, the vector locations that provide the lowest standard deviation in the window,
at least below the provided \code{stdlevel}.}
\item{average}{The calculated average for the vector locations.}
\item{standarddeviation}{The calculated standard deviation for the vector locations.}
}
If no window below \code{stdlevel} is found, or if the desired average value is not found, or the closest
desired average value is not below \code{stdlevel}, \code{NULL} returns.
}
\description{
Within a numeric vector, find the locations of a window for which the values can be considered
constant according to the provided standard deviation level settings.
}
\seealso{
Other vector functions: 
\code{\link{dorollapply}()},
\code{\link{extendvector}()},
\code{\link{findconstantchunks}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{replaceelements}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()},
\code{\link{shift}()}
}
\concept{vector functions}
