% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedVector.R
\name{findconstantchunks}
\alias{findconstantchunks}
\title{Find constant value chunks}
\usage{
findconstantchunks(thevector, minpoints = 20, stdlevel = 0.05)
}
\arguments{
\item{thevector}{Numeric vector to search in.}

\item{minpoints}{Length of the window that requires to be constant. Default=\code{20}.}

\item{stdlevel}{The maximum standard deviation of the window. Any standard
deviation above this level will be considered not constant. Default=\code{0.05}.}
}
\value{
A data.table with one row for each chunk, with the following column names:
\describe{
\item{chunksize}{The amount of points in \code{thevector} that refers to this chunk.}
\item{chunkstartloc}{The very first location of \code{thevector} that refers to the start of
this window chunk.}
\item{chunkendloc}{The very last location of \code{thevector} that refers to the end of
this window chunk.}
\item{chunkaverage}{The average value of this chunk.}
\item{chunkstandarddeviation}{The standard deviation of this chunk (denominator \emph{n}).}
\item{minstdstartloc}{The very first location of \code{thevector} that provide the lowest
standard deviation in the chunk calculated for a window of size \code{minpoints}.}
\item{minstdendloc}{The very last location of \code{thevector} that provide the lowest
standard deviation in the chunk calculated for a window of size \code{minpoints}.}
\item{minstdaverage}{The calculated average value for the vector locations that
provide the lowest standard deviation in the chunk.}
\item{minstdstandarddeviation}{The standard deviation for the vector locations
that provide the lowest standard deviation in the chunk (denominator \emph{n})
calculated for a window of size \code{minpoints}.}
}
If no window below \code{stdlevel} is found, \code{NULL} returns.
}
\description{
Within a numeric vector, find all locations of a window for which the values can be considered
constant according to the provided standard deviation level settings.
}
\details{
In contrast to the \code{\link{findconstant}} function, which either returns a constant window
chunk for a provided desired value or the 'most constant' chunk (based on the lowest standard
deviation), this function returns all possible chunks in a vector that can be considered constant.
}
\seealso{
Other vector functions: 
\code{\link{dorollapply}()},
\code{\link{extendvector}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{replaceelements}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()},
\code{\link{shift}()}
}
\concept{vector functions}
