% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedLists.R
\name{forcelisttype}
\alias{forcelisttype}
\title{Force all values to be of list type}
\usage{
forcelisttype(thelist)
}
\arguments{
\item{thelist}{The list to traverse.}
}
\value{
List in which all values have been put into a list.
}
\description{
Recursively go over the list and transform all values to be of list type.
}
\seealso{
Other list functions: 
\code{\link{leavenames}()},
\code{\link{listdepth}()},
\code{\link{stripnames}()}
}
\concept{list functions}
