% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedPlotting.R
\name{getdefaultlabeltext}
\alias{getdefaultlabeltext}
\title{Get label text}
\usage{
getdefaultlabeltext(plotdata, plotsettings, whichaxis = 0)
}
\arguments{
\item{plotdata}{The data to use in the plot.}

\item{plotsettings}{The settings to use in the plot (see \code{\link{getdefaultplotsettings}}).}

\item{whichaxis}{The integer value that indicates the y axis from the plotsettings is used.
Default=\code{0}, which does not specify a specific y axis.}
}
\value{
The label text for each data point.
}
\description{
Make the label text for each data point in a graph. This is the default function, other specialized
functions can be written for specific data. This function is provided as a variable in the plotsettings
used in plotting.
}
\seealso{
Other plot functions: 
\code{\link{getdefaulthovertext}()},
\code{\link{getdefaultplotsettings}()},
\code{\link{makeplotsettings}()},
\code{\link{plot_makecolors}()},
\code{\link{plot_makemarkers}()},
\code{\link{plotcandlestick}()},
\code{\link{plotdensityfunctions}()},
\code{\link{plotecdffunctions}()},
\code{\link{plothistogram}()},
\code{\link{plotpiechart}()},
\code{\link{plotscatterwithlink}()},
\code{\link{plotscatter}()},
\code{\link{sampledashboard}()},
\code{\link{updateplotsettings}()}
}
\concept{plot functions}
