% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedPlotting.R
\name{getdefaultplotsettings}
\alias{getdefaultplotsettings}
\title{Plot settings}
\usage{
getdefaultplotsettings()
}
\value{
List with default plot settings.
}
\description{
Get the default plotsettings for use in plotting functions. Settings contain the following elements:
\describe{
\item{\code{xaxis}}{This sets the column from the data.table to be used for the x axis.
A string containing only one column name.
Default=\code{"x"}.
}
\item{\code{xaxisrangemode}}{This sets the x axis range setting.
If \code{"normal"}, all data is displayed. If \code{"tozero"}, the zero axis
is always shown. If \code{"nonnegative"}, the x axis range never goes below zero.
Default=\code{"normal"}
}
\item{\code{yaxes}}{This sets the columns from the data.table to display on the y axis.
Can be more than one string containing the column names.
If \code{graphtype="candlestick"} this needs to be of at least length 4,
containing the column information for 'open', 'close', 'high', 'low', in that order. Any
additional columns will be added as a regular scatter trace.
If \code{graphtype="piechart"} only the first provided column is used.
If \code{graphtype="treemap"} only the first provided column is used.
Default=\code{c("y")}.
}
\item{\code{yaxesvisible}}{This sets the visibility settings for the y axes.
Must be of same length as \code{plotsettings$yaxes},
and must be strings containing either \code{YES}, \code{NO}, or \code{legendonly}.
Default=\code{c("YES")}.
}
\item{\code{yaxeslinecolors}}{This sets the main display colors for the y axes.
Must be of same length as \code{plotsettings$yaxes},
and must be strings containing a color indication.
If \code{graphtype="scatter"} this setting is used for line colors.
If \code{graphtype="scatterstack"} this setting is used for line colors.
Fill colors will be automatically made based on this color.
If \code{graphtype="histogram"} this setting is used for bar colors.
If \code{graphtype="candlestick"} this setting is omitted.
If \code{graphtype="piechart"} this setting is omitted.
If \code{graphtype="treemap"} this setting is omitted.
Default=\code{c("#6C9DDC")}.
}
\item{\code{yaxesmarkercolors}}{This sets the columns from the data.table to use
as data source for the color of the markers on the representative y axis.
Must be of same length as \code{plotsettings$yaxes}
and be more than one string containing the column names, \code{NA} where
no markers are to be displayed, or \code{"linecolor"} if the marker needs to take the
same color as the line.
Works together with \code{plotsettings$markercolorsettings},
where the data in those columns are associated with colors.
If \code{graphtype="scatter"} this setting is used for marker colors.
If \code{graphtype="scatterstack"} this setting is used for marker colors.
If \code{graphtype="histogram"} this setting is used omitted.
If \code{graphtype="candlestick"} this setting is omitted.
If \code{graphtype="piechart"} this setting is used for pie colors.
If \code{graphtype="treemap"} this setting is used for block colors.
Default=\code{c("linecolor")}.
}
\item{\code{yaxesmarkersymbols}}{This sets the columns from the data.table to use
as data source for the symbols of the markers on the representative y axis.
Must be of same length as \code{plotsettings$yaxes}.
Can be more than one string containing the column names, or \code{NA} where
no symbols are to be displayed. Works together with \code{plotsettings$markersymbolsettings},
where the specific symbols for columns are associated.
If \code{graphtype="scatter"} this setting is used for marker colors.
If \code{graphtype="scatterstack"} this setting is used for marker colors.
If \code{graphtype="histogram"} this setting is used omitted.
If \code{graphtype="candlestick"} this setting is omitted.
If \code{graphtype="piechart"} this setting is omitted.
If \code{graphtype="treemap"} this setting is omitted.
Default=\code{c(NA)}.
}
\item{\code{yaxesmarkersizes}}{This sets the size of the marker (if displayed). Default=\code{8}.
Can be a single value, making all markers equal in size, or the same length as \code{plotsettings$yaxes}
allowing each y axis to have a different size marker. Missing values will use the first
provided marker size.
}
\item{\code{yaxesnames}}{This sets the name of the y axes, as to be displayed in the
legend. Also allows legend grouping by name. Must be of same length as \code{plotsettings$yaxes},
and must be strings containing the name or \code{NA} to use the same name as
provided in \code{plotsettings$yaxes}.
Default=\code{c(NA)}.
}
\item{\code{yaxisrangemode}}{This sets the y axis range setting.
If \code{"normal"}, all data is displayed.
If \code{"tozero"}, the zero axis is always shown.
If \code{"nonnegative"}, the y axis range never goes below zero.
Default=\code{"normal"}. This setting works globally, and not for each y axis individually.
}
\item{\code{displaytitle}}{This sets the graph title to a specified text. Must be a string
or \code{NULL} to display no title. Default=\code{NULL}.
}
\item{\code{displaylegend}}{This sets the visibility of a legend in the graph
(either \code{TRUE} or \code{FALSE}). Default=\code{TRUE}.
}
\item{\code{graphtype}}{This sets the graph type. Currently supported are
\code{"scatter"}  (will call \code{\link{plotscatter}}),
\code{"scatterstack"}  (will call \code{\link{plotscatter}}),
\code{"histogram"} (will call \code{\link{plothistogram}}),
\code{"candlestick"} (will call \code{\link{plotcandlestick}}),
\code{"piechart"} (will call \code{\link{plotpiechart}}), and
\code{"treemap"} (will call \code{\link{plottreemap}}).
Checks for \code{graphtype} happen in the dashboard, not in the \code{plot-} functions.
Default=\code{"scatter"}.
}
\item{\code{graphmode}}{This sets the mode of the graph, and is dependent
on what \code{graphtype} is chosen. Default=\code{NA}.
If \code{graphtype="scatter"} this setting can be a vector with the same length as
\code{yaxes} and will be either \code{"lines"}, \code{"lines+markers"} or \code{"markers"}
for the specific y axis data to be plotted in that mode. If the length is insufficient,
it will default to the \code{NA} setting for that y axis.
When \code{NA} is provided, it will use the \code{"lines+markers"} setting.
If \code{graphtype="scatterstack"} this setting is the same as \code{graphtype="scatter"}
but \code{"none"} can be added to hide the lines.
When \code{NA} is provided, it will use the \code{"none"} setting.
If \code{graphtype="histogram"} this setting is used for the \code{barmode} setting of
the plot layout. Can be either \code{"group"}, \code{"stack"} or \code{"overlay"}.
When \code{NA} is provided, it will use the \code{"group"} setting.
If \code{graphtype="candlestick"} this setting can be a vector with the same length as
\code{yaxes} and will be either \code{"lines"}, \code{"lines+markers"} or \code{"markers"}
for the specific y axis data to be plotted in that mode. If the length is insufficient,
it will default to the \code{NA} setting for that y axis.
When \code{NA} is provided, it will use the \code{"lines+markers"} setting.
If \code{graphtype="piechart"} this setting is omitted.
If \code{graphtype="treemap"} this setting is omitted.
}
\item{\code{markercolorsettings}}{This associates the column names of the data.table with the content in that column, and the color to use.
The association of data with colors only happens when \code{yaxesmarkercolors} is provided.
Consists of a list of list, in which the interior list contains:
\describe{
\item{columnnames}{String vector of data.table column name that contains the data to be mapped.}
\item{columncontent}{Vector of discrete data points to be found in this column of the data.table.
If only two values are provided and they are numeric, then they are assumed to be
the minimum and maximum value for a color ramp to be generated.
Leaving this list item empty (NULL or NA) then a color ramp is automatically assumed, mapping
the minimum and maximum value of the data.}
\item{columncolors}{The color to use for each specific data point, or the various colors
(minimum 2) of the color ramp to generate. To map discrete data points this list item must be of the same
length as \code{columncontent} and can be a color name, a hex color,
or \code{"linecolor"} (which takes the same color as the line,  specified in \code{plotsettings$yaxeslinecolors}).}
}
The default is a list too long to report here, but can be found by simply executing this function.
}
\item{\code{markersymbolsettings}}{This associates the column names of the data.table with the content in that column, and the symbol to use.
The association of data with colors only happens when \code{yaxesmarkersymbols} is provided.
Consists of a list of list, in which the interior list contains:
\describe{
\item{columnnames}{String vector of data.table column name that contains the data to be mapped.}
\item{columncontent}{Vector of discrete data points to be found in this column of the data.table.}
\item{columnsymbols}{The symbol to use for each specific data point. Must be of the same length as \code{columncontent}.}
}
}
\item{\code{datafix_keepmissingvalues}}{This informs the plotting functions what to do with missing values. If \code{TRUE},
missing values are kept and can cause gaps in the display. If \code{FALSE}, missing values are removed which will create a
continuous display. Default=\code{TRUE}.
}
\item{\code{function_hovertext}}{This informs the plotting functions which function to use to create the hover text format.
Set to \code{NULL} or to a function that returns \code{NULL} to not display any hover text.
Default=\code{getdefaulthovertext}.
}
\item{\code{function_labeltext}}{This informs the plotting functions which function to use to create the label text format.
Set to \code{NULL} or to a function that returns \code{NULL} to not display any label text.
Default=\code{getdefaultlabeltext}, which simply returns \code{NULL} to avoid default label display.
}
\item{\code{labelposition}}{This sets the positioning method of the label text.
If \code{graphtype="scatter"} or \code{graphtype="scatterstack"} or \code{graphtype="treemap"} this setting can be a combination
of \code{"top"}, \code{ "middle"} or \code{"bottom"} with \code{"left"}, \code{"center"} or \code{"right"}.
If \code{graphtype="histogram"} or \code{graphtype="piechart"} this setting can be either of \code{"inside"},
\code{"outside"}, \code{"auto"} or \code{"none"}.
If \code{graphtype="candlestick"} this setting is only used for the additional scatter traces that follow the candlestick traces.
Default=\code{NA}, which takes either \code{"middle center"} or \code{"auto"} depending on \code{graphtype}.
}
}
}
\seealso{
Other plot functions: 
\code{\link{getdefaulthovertext}()},
\code{\link{getdefaultlabeltext}()},
\code{\link{makeplotsettings}()},
\code{\link{plot_makecolors}()},
\code{\link{plot_makemarkers}()},
\code{\link{plotcandlestick}()},
\code{\link{plotdensityfunctions}()},
\code{\link{plotecdffunctions}()},
\code{\link{plothistogram}()},
\code{\link{plotpiechart}()},
\code{\link{plotscatterwithlink}()},
\code{\link{plotscatter}()},
\code{\link{sampledashboard}()},
\code{\link{updateplotsettings}()}
}
\concept{plot functions}
