% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedFile.R
\name{getfileextensions}
\alias{getfileextensions}
\title{Get file extension}
\usage{
getfileextensions(fullpaths)
}
\arguments{
\item{fullpaths}{The full paths or file name to get extensions from.}
}
\value{
List of extensions.
}
\description{
Get the file extension from a full path or filename.
}
\seealso{
Other file functions: 
\code{\link{getfilenames}()},
\code{\link{getfiletable}()}
}
\concept{file functions}
