% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedFile.R
\name{getfilenames}
\alias{getfilenames}
\title{Get file name}
\usage{
getfilenames(fullpaths, excludeextension = FALSE)
}
\arguments{
\item{fullpaths}{The full paths to get file names from.}

\item{excludeextension}{Set to \code{TRUE} to exclude the extension from the returned filenames..
Default=\code{FALSE}, which includes the extension.}
}
\value{
List of file names.
}
\description{
Get the file name from a full path.
}
\seealso{
Other file functions: 
\code{\link{getfileextensions}()},
\code{\link{getfiletable}()}
}
\concept{file functions}
