% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedFile.R
\name{getfiletable}
\alias{getfiletable}
\title{Get directory file list}
\usage{
getfiletable(
  dirtocheck,
  fileextension = NULL,
  includesubdirs = FALSE,
  printmessage = FALSE
)
}
\arguments{
\item{dirtocheck}{The directory to obtain the files from.}

\item{fileextension}{One of more strings specifying the extensions (without dot, such as \code{"zip"}) of the files to return.
Default=\code{NULL}, which returns all files.}

\item{includesubdirs}{Set to \code{TRUE} to also include the files in all subdirectories.
Default=\code{FALSE}, which only checks the cirrent directory.}

\item{printmessage}{Set to \code{TRUE} to display the amount of files found in the directory in a message.
Default=\code{FALSE}, which does not print a message.}
}
\value{
Data.table with the files, containing the columns FULLPATH (full non-relative path), DIRONLY (directory of the file),
FILENAME (file name without extension), and EXTENSION (file extension).
}
\description{
Get a list of files in a specific directory.
}
\seealso{
Other file functions: 
\code{\link{getfileextensions}()},
\code{\link{getfilenames}()}
}
\concept{file functions}
