% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedPlotting.R
\name{gethuecolors}
\alias{gethuecolors}
\title{Generate hue colors}
\usage{
gethuecolors(n)
}
\arguments{
\item{n}{The amount of colors to generate}
}
\value{
Vector containing n amount of colors
}
\description{
Generate a range of hues (colors) for graphs
}
\seealso{
Other plot assists: 
\code{\link{getpalettecolors}()}
}
\concept{plot assists}
