% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedPlotting.R
\name{getpalettecolors}
\alias{getpalettecolors}
\title{Generate palette colors}
\usage{
getpalettecolors(observations)
}
\arguments{
\item{observations}{character vector or number vector}
}
\value{
Data.table with the columns \code{OBSERVATIONS} and \code{COLOR}.
}
\description{
Generate a range of palette based colors. Uses the 70 unique colors from the RColorBrewer
library and recycles them to match the number of observations.
}
\examples{
observations =  1:100
DT_test = merge(
    x=data.table(BLABLA = observations),
    y=ut_create_colors(observations),
    by.x="BLABLA",
    by.y="OBSERVATIONS"
  )

  table(
    data.table(
      table(DT_test$COLOR)
    )[order(-N)]$N
  )
}
\seealso{
Other plot assists: 
\code{\link{gethuecolors}()}
}
\concept{plot assists}
