% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedConversion.R
\name{hextobinary}
\alias{hextobinary}
\title{Hex to binary}
\usage{
hextobinary(hexvalues)
}
\arguments{
\item{hexvalues}{String vector containing hexadecimal values.}
}
\value{
String vector containing the related binary bits, or NA where the provided
string is invalid (not a hex value representation).
}
\description{
Transform hexadecimal values (in string vector) to binary bits (also in string vector)
}
\examples{
hexdat <- replicate(10, paste(format.hexmode(sample(16,4)-1),collapse=''))
hextobinary(hexdat)
}
\seealso{
Other conversion functions: 
\code{\link{bitsinbinary}()},
\code{\link{convertcoords_degrees2decimal}()}
}
\concept{conversion functions}
