% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedMath.R
\name{inbetween}
\alias{inbetween}
\title{In between}
\usage{
inbetween(value, minval, maxval)
}
\arguments{
\item{value}{The value to check}

\item{minval}{The minimum value of the range}

\item{maxval}{The maximum value of the range}
}
\value{
\code{TRUE} if the value is in between the range (minimum and maximum
values included), \code{FALSE} if this is not the case.
}
\description{
Check if a value is in between minimum and maximum value
}
