% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{innerjoin}
\alias{innerjoin}
\title{Inner join data.tables}
\usage{
innerjoin(
  basetable,
  additionaltable,
  basetablejoincolumns,
  additionaltablejoincolumns = NULL,
  additionaltableextractcolumns = NULL,
  removejoincols = NA
)
}
\arguments{
\item{basetable}{The base data.table.}

\item{additionaltable}{The additional data.table.}

\item{basetablejoincolumns}{The column name string of the base data.table to join on.}

\item{additionaltablejoincolumns}{The column name string of the additional data.table to join on.
Default=\code{NULL}, which assumes the same column name as \code{basetablejoincolumns}.}

\item{additionaltableextractcolumns}{The column name strings of the additional data.table to add to the base data.table.
Default=\code{NULL}, which assumes that all columns need to be joined.}

\item{removejoincols}{If \code{TRUE}, it will remove the provided \code{basetablejoincolumns} from the resulting table.
If \code{FALSE}, it will keep the provided \code{basetablejoincolumns} in the result. If \code{NA} (default), it will check if
\code{basetablejoincolumns} and \code{additionaltablejoincolumns} are the same. The column names that are different will result
in the respective \code{basetablejoincolumns} to be removed.

For example, let's take \code{basetablejoincolumns=c("ID","SECID")} and \code{additionaltablejoincolumns=c("ID","ID2")}.
With \code{removejoincols=TRUE}, the column \code{"ID2"} will return in the result, but columns \code{"ID"} and \code{"SECID"} will not.
With \code{removejoincols=FALSE}, the columns \code{"ID.x"}, \code{"ID.y"}, \code{"ID2"} and \code{"SECID"} will return.
Note that column \code{ID} exists in both the base as well as additional table, which results in the \code{"ID.x"} and \code{"ID.y"} columns.
With \code{removejoincols=NA} the column \code{"ID"} and \code{"ID2"} will return in the result, but column \code{"SECID"} will not.}
}
\value{
The joined data.table or, when join column(s) or extracted column(s) don't exist, the original base data.table.
}
\description{
Execute an inner join on two data.tables, meaning that only data present in both data.tables will be kept.
}
\seealso{
Other data.table functions: 
\code{\link{addcolumn}()},
\code{\link{addisoinfo}()},
\code{\link{copytoclipboard}()},
\code{\link{dojoin}()},
\code{\link{extractcategories}()},
\code{\link{fulljoin}()},
\code{\link{getcolumnclasses}()},
\code{\link{leftjoin}()},
\code{\link{loadlargetable}()},
\code{\link{loadsastable}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{makeintegercolumns}()},
\code{\link{makenumericcolumns}()},
\code{\link{mapcolumnclasses}()},
\code{\link{prefixcolnames}()},
\code{\link{removeNAfromtable}()},
\code{\link{removeallNAcolumns}()},
\code{\link{removeallNArows}()},
\code{\link{removecolumns}()},
\code{\link{removeisoinfo}()},
\code{\link{removerows}()},
\code{\link{renamecolumns}()},
\code{\link{rightjoin}()},
\code{\link{runcolfun}()},
\code{\link{selectuniques}()},
\code{\link{selectwherein}()},
\code{\link{selectwhere}()},
\code{\link{setcolclass}()},
\code{\link{setcolnames}()},
\code{\link{setcoltofront}()},
\code{\link{setcolumns}()},
\code{\link{unlistcolumns}()},
\code{\link{updatetablevalues}()}
}
\concept{data.table functions}
