% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedVector.R
\name{intersect_all}
\alias{intersect_all}
\title{Get the same vector elements}
\usage{
intersect_all(a, b, ...)
}
\arguments{
\item{a}{First vector}

\item{b}{Second vector}

\item{...}{any further vectors}
}
\value{
Vector containing elements that were in all provided vectors.
}
\description{
For multiple vectors, get the vector that contains the same values that are within all provided vectors.
}
