% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedGeneral.R
\docType{package}
\name{JuicedGeneral-package}
\alias{JuicedGeneral}
\alias{JuicedGeneral-package}
\title{JuicedGeneral}
\description{
These are some general useful functions that have not been included at any other location. Quite varied from calculations to plotting.
}
\author{
\strong{Maintainer}: Joost Kuckartz \email{jkuckartz1984@hotmail.com} [copyright holder]

}
