% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedLists.R
\name{leavenames}
\alias{leavenames}
\title{Leave list with only names}
\usage{
leavenames(thelist, thenames)
}
\arguments{
\item{thelist}{List to only leave the provided names in.}

\item{thenames}{Vector of strings containing the names of the elements to leave in the list.}
}
\value{
List with only the named elements remaining.
}
\description{
Removes all named elements except the provided names from a list. If none of the
provided names are found within the list, the full list is kept (therefore keeping
the list structure).
Executes recursively so that it also works within lists in lists.
}
\seealso{
Other list functions: 
\code{\link{forcelisttype}()},
\code{\link{listdepth}()},
\code{\link{stripnames}()}
}
\concept{list functions}
