% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedMath.R
\name{linearcalc}
\alias{linearcalc}
\title{Linear interpolation with normalization}
\usage{
linearcalc(inputvector, minval, maxval, inverted = FALSE)
}
\arguments{
\item{inputvector}{The values to calculate on.}

\item{minval}{The minimum value.}

\item{maxval}{The maximum value.}

\item{inverted}{Set to \code{TRUE} to invert the result, so that the maximum value and above represent 0 and the minimum value and below represent 1
(default=\code{FALSE}, which uses the minimum value and below to represent 0 and the maximum value and above to represent 1).}
}
\description{
Perform linear interpolation and normalization in between a minimum and maximum value.
The minimum value and below represent 0, the maximum value and above represent 1, unless \code{inverted=TRUE}, which reverses
this logic.
}
\seealso{
Other predictors: 
\code{\link{linearextrapolate}()}
}
\concept{predictors}
