% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedMath.R
\name{linearextrapolate}
\alias{linearextrapolate}
\title{Linear extrapolation}
\usage{
linearextrapolate(x, y, xi)
}
\arguments{
\item{x}{The values on the x axis.}

\item{y}{Corresponding values on the y axis.}

\item{xi}{The extrapolated x axis value to calculate a new y value for.}
}
\value{
The calculated extrapolated y value.
}
\description{
Linear extrapolation
}
\seealso{
Other predictors: 
\code{\link{linearcalc}()}
}
\concept{predictors}
