% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedLists.R
\name{listdepth}
\alias{listdepth}
\title{Find list depth}
\usage{
listdepth(thelist, thisdepth = 0)
}
\arguments{
\item{thelist}{List to find the depth for.}

\item{thisdepth}{Used for recursion to find the depth but can be modified
if you provide a list within list already. Default=\code{0}.}
}
\value{
Integer that identifies the depth of a list.
}
\description{
Find the depth of a list, which is an integer value of the
furthest recursion into the lists in lists.
}
\seealso{
Other list functions: 
\code{\link{forcelisttype}()},
\code{\link{leavenames}()},
\code{\link{stripnames}()}
}
\concept{list functions}
