% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedLoading.R
\name{loaddependentpackages}
\alias{loaddependentpackages}
\title{Load dependencies}
\usage{
loaddependentpackages(dir)
}
\arguments{
\item{dir}{The directory of the development package.}
}
\description{
For a development package, only load the dependencies.
}
\seealso{
Other loading functions: 
\code{\link{loaddevelopmentlibrary}()},
\code{\link{loadlargetable}()},
\code{\link{loadlibrary}()},
\code{\link{loadsastable}()},
\code{\link{loadsource}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()}
}
\concept{loading functions}
