% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedLoading.R
\name{loaddevelopmentlibrary}
\alias{loaddevelopmentlibrary}
\title{Load a development library}
\usage{
loaddevelopmentlibrary(x, musthaveversion = "0.0.0")
}
\arguments{
\item{x}{Character string of the library to load.}

\item{musthaveversion}{The version that the package must have (default="0.0.0")}
}
\description{
Load a specific named development library.
}
\seealso{
Other loading functions: 
\code{\link{loaddependentpackages}()},
\code{\link{loadlargetable}()},
\code{\link{loadlibrary}()},
\code{\link{loadsastable}()},
\code{\link{loadsource}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()}
}
\concept{loading functions}
