% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{loadlargetable}
\alias{loadlargetable}
\title{Load a large table}
\usage{
loadlargetable(dir, fname, forcesas = FALSE)
}
\arguments{
\item{dir}{Directory in which the file can be found. Must end with '/'.}

\item{fname}{File name of the file (without extension).}

\item{forcesas}{Set to TRUE if the original SAS table needs to be loaded (it otherwise uses the quicker rds file).}
}
\value{
Data.table on success, NULL on failure.
}
\description{
Load a large SAS or RDS data table from file, with an automatic RDS existance check,
and return it as a data.table. If a SAS table is loaded, it will automatically save
it as RDS in the same directory for future faster loading.
}
\details{
Depending on the size of the SAS table, this function can take a while to finish.
It does not have a progress indicator.
}
\seealso{
Other loading functions: 
\code{\link{loaddependentpackages}()},
\code{\link{loaddevelopmentlibrary}()},
\code{\link{loadlibrary}()},
\code{\link{loadsastable}()},
\code{\link{loadsource}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()}

Other data.table functions: 
\code{\link{addcolumn}()},
\code{\link{addisoinfo}()},
\code{\link{copytoclipboard}()},
\code{\link{dojoin}()},
\code{\link{extractcategories}()},
\code{\link{fulljoin}()},
\code{\link{getcolumnclasses}()},
\code{\link{innerjoin}()},
\code{\link{leftjoin}()},
\code{\link{loadsastable}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{makeintegercolumns}()},
\code{\link{makenumericcolumns}()},
\code{\link{mapcolumnclasses}()},
\code{\link{prefixcolnames}()},
\code{\link{removeNAfromtable}()},
\code{\link{removeallNAcolumns}()},
\code{\link{removeallNArows}()},
\code{\link{removecolumns}()},
\code{\link{removeisoinfo}()},
\code{\link{removerows}()},
\code{\link{renamecolumns}()},
\code{\link{rightjoin}()},
\code{\link{runcolfun}()},
\code{\link{selectuniques}()},
\code{\link{selectwherein}()},
\code{\link{selectwhere}()},
\code{\link{setcolclass}()},
\code{\link{setcolnames}()},
\code{\link{setcoltofront}()},
\code{\link{setcolumns}()},
\code{\link{unlistcolumns}()},
\code{\link{updatetablevalues}()}
}
\concept{data.table functions}
\concept{loading functions}
