% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedLoading.R
\name{loadlibrary}
\alias{loadlibrary}
\title{Load a library}
\usage{
loadlibrary(x)
}
\arguments{
\item{x}{Character string of the library to load.}
}
\description{
Load a specific named library. If it is not yet installed, it will install it.
}
\seealso{
Other loading functions: 
\code{\link{loaddependentpackages}()},
\code{\link{loaddevelopmentlibrary}()},
\code{\link{loadlargetable}()},
\code{\link{loadsastable}()},
\code{\link{loadsource}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()}
}
\concept{loading functions}
