% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{loadsastable}
\alias{loadsastable}
\title{Load a SAS data table}
\usage{
loadsastable(dir, fname)
}
\arguments{
\item{dir}{Directory in which the file can be found. Must end with '/'.}

\item{fname}{File name of the file (without extension).}
}
\value{
Data.table on success, NULL on failure.
}
\description{
Load a SAS data table from file and return it as a data.table.
}
\details{
Depending on the size of the SAS table, this function can take a while to finish. It does not have a progress indicator.
}
\seealso{
Other loading functions: 
\code{\link{loaddependentpackages}()},
\code{\link{loaddevelopmentlibrary}()},
\code{\link{loadlargetable}()},
\code{\link{loadlibrary}()},
\code{\link{loadsource}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()}

Other data.table functions: 
\code{\link{addcolumn}()},
\code{\link{addisoinfo}()},
\code{\link{copytoclipboard}()},
\code{\link{dojoin}()},
\code{\link{extractcategories}()},
\code{\link{fulljoin}()},
\code{\link{getcolumnclasses}()},
\code{\link{innerjoin}()},
\code{\link{leftjoin}()},
\code{\link{loadlargetable}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{makeintegercolumns}()},
\code{\link{makenumericcolumns}()},
\code{\link{mapcolumnclasses}()},
\code{\link{prefixcolnames}()},
\code{\link{removeNAfromtable}()},
\code{\link{removeallNAcolumns}()},
\code{\link{removeallNArows}()},
\code{\link{removecolumns}()},
\code{\link{removeisoinfo}()},
\code{\link{removerows}()},
\code{\link{renamecolumns}()},
\code{\link{rightjoin}()},
\code{\link{runcolfun}()},
\code{\link{selectuniques}()},
\code{\link{selectwherein}()},
\code{\link{selectwhere}()},
\code{\link{setcolclass}()},
\code{\link{setcolnames}()},
\code{\link{setcoltofront}()},
\code{\link{setcolumns}()},
\code{\link{unlistcolumns}()},
\code{\link{updatetablevalues}()}
}
\concept{data.table functions}
\concept{loading functions}
