% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedLoading.R
\name{loadsource}
\alias{loadsource}
\title{Load source files}
\usage{
loadsource(dir, openeditor = FALSE, loaddependencies = FALSE)
}
\arguments{
\item{dir}{The directory to load the source files from.}

\item{openeditor}{Set to \code{TRUE} to open the source files in the editor. In RStudio, this allows debugging
of the functions through automatic error breakpoints or using the F2 function lookup shortcut followed by direct
editing. Default=\code{FALSE}, which only sources the files without opening an editor.}

\item{loaddependencies}{Set to \code{TRUE} to load all libraries that the source package depends on. This allows
for the loading of the source plus all dependent libraries, without actually calling \code{library} of the actual
package. This allows for setting breakpoints and debugging the opened source code without RStudio complaining that
an updated package is needed. Default=\code{FALSE}, which does not call dependent libraries.}
}
\description{
Load all source files in a specific directory. Can be used to source a package, allowing direct package
editing while running it. Of course, after code modification the package needs to be built and installed
again.
}
\seealso{
Other loading functions: 
\code{\link{loaddependentpackages}()},
\code{\link{loaddevelopmentlibrary}()},
\code{\link{loadlargetable}()},
\code{\link{loadlibrary}()},
\code{\link{loadsastable}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()}
}
\concept{loading functions}
