% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{loadspecifictable}
\alias{loadspecifictable}
\title{Load a RDS, CSV, XLSX or SAS file}
\usage{
loadspecifictable(dir, fname, filetype = "RDS", largetable = FALSE)
}
\arguments{
\item{dir}{Directory in which the file can be found. If it does not end with '/', this directory separator will be added.}

\item{fname}{File name of the file (without extension).}

\item{filetype}{The file type of the file to force load ("RDS", "CSV", "XLSX", or "SAS").
Default = "RDS". Setting this to a either "CSV", "XLSX" or "SAS" forces a reload.}

\item{largetable}{Set to \code{TRUE} if this is a large table, which stores the table to RDS after loading.
Default=\code{FALSE}, which does not store the result in a RDS file.}
}
\value{
Data.table on success, NULL on failure.
}
\description{
Loads the file and returns it as a data.table. The loading priority follows RDS, then CSV, then XLSX, then SAS.
}
\details{
Depending on the size of the file, this function can take a while to finish. It does not have a progress indicator.
}
\seealso{
Other loading functions: 
\code{\link{loaddependentpackages}()},
\code{\link{loaddevelopmentlibrary}()},
\code{\link{loadlargetable}()},
\code{\link{loadlibrary}()},
\code{\link{loadsastable}()},
\code{\link{loadsource}()},
\code{\link{loadtable}()}

Other data.table functions: 
\code{\link{addcolumn}()},
\code{\link{addisoinfo}()},
\code{\link{copytoclipboard}()},
\code{\link{dojoin}()},
\code{\link{extractcategories}()},
\code{\link{fulljoin}()},
\code{\link{getcolumnclasses}()},
\code{\link{innerjoin}()},
\code{\link{leftjoin}()},
\code{\link{loadlargetable}()},
\code{\link{loadsastable}()},
\code{\link{loadtable}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{makeintegercolumns}()},
\code{\link{makenumericcolumns}()},
\code{\link{mapcolumnclasses}()},
\code{\link{prefixcolnames}()},
\code{\link{removeNAfromtable}()},
\code{\link{removeallNAcolumns}()},
\code{\link{removeallNArows}()},
\code{\link{removecolumns}()},
\code{\link{removeisoinfo}()},
\code{\link{removerows}()},
\code{\link{renamecolumns}()},
\code{\link{rightjoin}()},
\code{\link{runcolfun}()},
\code{\link{selectuniques}()},
\code{\link{selectwherein}()},
\code{\link{selectwhere}()},
\code{\link{setcolclass}()},
\code{\link{setcolnames}()},
\code{\link{setcoltofront}()},
\code{\link{setcolumns}()},
\code{\link{unlistcolumns}()},
\code{\link{updatetablevalues}()}
}
\concept{data.table functions}
\concept{loading functions}
