% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{loadtable}
\alias{loadtable}
\title{Load a tabular file}
\usage{
loadtable(
  dir,
  fname,
  filetype = "RDS",
  quickstore = TRUE,
  forceallsheets = FALSE,
  ...
)
}
\arguments{
\item{dir}{Directory in which the file can be found. If it does not end with '/', this directory separator will be added.}

\item{fname}{File name of the file (without extension).}

\item{quickstore}{Set to \code{FALSE} to NOT store the data to .RDS after loading.
Default=\code{TRUE}, which does store an RDS file that makes future loading faster.}

\item{forceallsheets}{Set to \code{TRUE} if all sheets of an .XLSX file need to be loaded, without asking for user input.}
}
\value{
Data.table on success, \code{NULL} on failure. It will return a list of data.tables
in case an .xlsx with multiple sheets is selected to read.
}
\description{
Loads a tabular file and returns it as a data.table. Automatically loads the most efficient type first if available.
It can read the following file types:
RDS, CSV, XLSX, SAS
}
\seealso{
Other loading functions: 
\code{\link{loaddependentpackages}()},
\code{\link{loaddevelopmentlibrary}()},
\code{\link{loadlargetable}()},
\code{\link{loadlibrary}()},
\code{\link{loadsastable}()},
\code{\link{loadsource}()},
\code{\link{loadspecifictable}()}

Other data.table functions: 
\code{\link{addcolumn}()},
\code{\link{addisoinfo}()},
\code{\link{copytoclipboard}()},
\code{\link{dojoin}()},
\code{\link{extractcategories}()},
\code{\link{fulljoin}()},
\code{\link{getcolumnclasses}()},
\code{\link{innerjoin}()},
\code{\link{leftjoin}()},
\code{\link{loadlargetable}()},
\code{\link{loadsastable}()},
\code{\link{loadspecifictable}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{makeintegercolumns}()},
\code{\link{makenumericcolumns}()},
\code{\link{mapcolumnclasses}()},
\code{\link{prefixcolnames}()},
\code{\link{removeNAfromtable}()},
\code{\link{removeallNAcolumns}()},
\code{\link{removeallNArows}()},
\code{\link{removecolumns}()},
\code{\link{removeisoinfo}()},
\code{\link{removerows}()},
\code{\link{renamecolumns}()},
\code{\link{rightjoin}()},
\code{\link{runcolfun}()},
\code{\link{selectuniques}()},
\code{\link{selectwherein}()},
\code{\link{selectwhere}()},
\code{\link{setcolclass}()},
\code{\link{setcolnames}()},
\code{\link{setcoltofront}()},
\code{\link{setcolumns}()},
\code{\link{unlistcolumns}()},
\code{\link{updatetablevalues}()}
}
\concept{data.table functions}
\concept{loading functions}
