% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedData.R
\name{localmaxima}
\alias{localmaxima}
\title{Get local maxima}
\usage{
localmaxima(x)
}
\arguments{
\item{x}{Vector of values to get the local maxima locations of.}
}
\value{
The vector locations (like the \code{which} function) where
local maxima occur.
}
\description{
Get local maxima
}
\seealso{
Other vector functions: 
\code{\link{dorollapply}()},
\code{\link{extendvector}()},
\code{\link{findconstantchunks}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{replaceelements}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()},
\code{\link{shift}()}

Other data functions: 
\code{\link{createbins}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{removeoffset}()},
\code{\link{timeseriesstats}()}
}
\concept{data functions}
\concept{vector functions}
