% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedData.R
\name{localminima}
\alias{localminima}
\title{Get local minima}
\usage{
localminima(x)
}
\arguments{
\item{x}{Vector of values to get the local minima locations of.}
}
\value{
The vector locations (like the \code{which} function) where
local minima occur.
}
\description{
Get local minima
}
\seealso{
Other vector functions: 
\code{\link{chunk}()},
\code{\link{dorollapply}()},
\code{\link{extendvector}()},
\code{\link{findconstantchunks}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{replaceelements}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()},
\code{\link{shift}()}

Other data functions: 
\code{\link{createbins}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{removeoffset}()},
\code{\link{timeseriesstats}()}
}
\concept{data functions}
\concept{vector functions}
