% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedData.R
\name{localminimamaxima}
\alias{localminimamaxima}
\title{Get local minima and maxima}
\usage{
localminimamaxima(x, threshold = 1)
}
\arguments{
\item{x}{Vector of values to get the local minima and maxima locations of.}

\item{threshold}{Integer threshold value to get wider spaced minima and maxima
the higher threshold is. Default = 1.}
}
\value{
Named list with vector locations (like the \code{which} function) where
(local) minima and  maxima occur.
}
\description{
Get local minima and maxima
}
\seealso{
Other vector functions: 
\code{\link{chunk}()},
\code{\link{dorollapply}()},
\code{\link{extendvector}()},
\code{\link{findconstantchunks}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{replaceelements}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()},
\code{\link{shift}()}

Other data functions: 
\code{\link{createbins}()},
\code{\link{localmaxima}()},
\code{\link{localminima}()},
\code{\link{removeoffset}()},
\code{\link{timeseriesstats}()}
}
\concept{data functions}
\concept{vector functions}
