% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDateTime.R
\name{makeconsistentdateformat}
\alias{makeconsistentdateformat}
\title{Fix dates}
\usage{
makeconsistentdateformat(thedates, asstring = NULL, timezone = "")
}
\arguments{
\item{thedates}{Vector with dates to convert. Can work with POSIXt, Date, numeric and string.
If numeric is provided, it assumes Unix timestamp in seconds. If \code{NULL} is provided it returns \code{NULL} too.}

\item{asstring}{A string format (following \code{\link[base]{strptime}}) to convert the date to. Default=\code{NULL},
which returns the date in POSIXct format.}

\item{timezone}{String determining the timezone (from \code{as.POSIXct}). Default=\code{""},
which uses the local time zone.}
}
\value{
The dates in POSIXct format with the provided timezone, or as a string through the provided format,
or \code{NULL} if the provided input is incorrect.
}
\description{
Make a consistent date format from a provided variable. Has internal class detection so
can be called on already properly formatted dates (POSIXt) too.
For date and time use \code{\link{makeconsistentdatetimeformat}} function.
}
\seealso{
Other date and time functions: 
\code{\link{addisoinfo}()},
\code{\link{findperiodcontained}()},
\code{\link{findperiodoverlapping}()},
\code{\link{makeconsistentdatetimeformat}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{removeisoinfo}()},
\code{\link{seqdays}()}
}
\concept{date and time functions}
