% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{makecoordinatecolumns}
\alias{makecoordinatecolumns}
\title{Make columns GPS coordinates}
\usage{
makecoordinatecolumns(
  thedata,
  thecolnames = NULL,
  degreestodecimal = FALSE,
  divisionfactor = 1
)
}
\arguments{
\item{thedata}{The data.table to modify.}

\item{thecolnames}{Column names to modify to numeric GPS coordinates. If commas are present,
they are replaced by dots (to represent decimals). Default=\code{NULL},
which transforms all columns that it detects to be of class character.}

\item{degreestodecimal}{Set to \code{TRUE} to assume the existing coordinates are
in degrees, and need to be converted to decimals. Default=\code{FALSE}, which assumes the
existing coordinates to be decimals already.}

\item{divisionfactor}{Provide a value to divide the coordinates by. In certain situations,
the provided coordinates have been multiplied by this factor to be stored more efficiently.
Default=\code{1}, which assumes the coordinates have been stored without multiplication.}
}
\value{
Returns the modified data.table invisibly (modified by reference).
}
\description{
Make specific columns from a data.table numeric GPS coordinates, if they are of character class.
}
\concept{data fixes}
