% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedPlotting.R
\name{makeplotsettings}
\alias{makeplotsettings}
\title{Make plotsettings}
\usage{
makeplotsettings(yaxesnames, sethuecolors = TRUE)
}
\arguments{
\item{yaxesnames}{Vector with strings representing columns to display on the y axes.}

\item{sethuecolors}{Set to \code{FALSE} to keep each y axis in the default color.
Default=\code{TRUE}, which makes a unique line color for each y axis.}
}
\value{
List with default plot settings as returned from \code{\link{getdefaultplotsettings}},
but with the default settings applied to more than one y axis.
}
\description{
Simplified function to create automatic plotsettings for a range of yaxes names.
}
\seealso{
Other plot functions: 
\code{\link{getdefaulthovertext}()},
\code{\link{getdefaultlabeltext}()},
\code{\link{getdefaultplotsettings}()},
\code{\link{plot_makecolors}()},
\code{\link{plot_makemarkers}()},
\code{\link{plotcandlestick}()},
\code{\link{plotdensityfunctions}()},
\code{\link{plotecdffunctions}()},
\code{\link{plothistogram}()},
\code{\link{plotpiechart}()},
\code{\link{plotscatterwithlink}()},
\code{\link{plotscatter}()},
\code{\link{sampledashboard}()},
\code{\link{updateplotsettings}()}
}
\concept{plot functions}
