% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{mapcolumnclasses}
\alias{mapcolumnclasses}
\title{Map column classes}
\usage{
mapcolumnclasses(thedata, thecolumns, theclasses, ...)
}
\arguments{
\item{thedata}{The data.table to modify.}

\item{thecolumns}{Vector string of column names of the data.table that
need to have a specific class.}

\item{theclasses}{Vector string of column classes that they need to be.
Need to be of same length as \code{thecolumns}. If the strings \code{"date"}
or \code{"datetime"} are provided but the columns are not yet in POSIXct,
the functions \code{\link{makedatecolumns}} and \code{\link{makedatetimecolumns}}
are called respectively.}

\item{...}{Additional parameters of the \code{\link{makedatecolumns}}
and \code{\link{makedatetimecolumns}} functions, if needed.}
}
\value{
Returns the modified data.table invisibly (modified by reference).
}
\description{
Ensure columns of a data.table have the right class and if not, transform those columns.
}
\seealso{
Other data.table functions: 
\code{\link{addcolumn}()},
\code{\link{addisoinfo}()},
\code{\link{copytoclipboard}()},
\code{\link{dojoin}()},
\code{\link{extractcategories}()},
\code{\link{fulljoin}()},
\code{\link{getcolumnclasses}()},
\code{\link{innerjoin}()},
\code{\link{leftjoin}()},
\code{\link{loadlargetable}()},
\code{\link{loadsastable}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{makeintegercolumns}()},
\code{\link{makenumericcolumns}()},
\code{\link{prefixcolnames}()},
\code{\link{removeNAfromtable}()},
\code{\link{removeallNAcolumns}()},
\code{\link{removeallNArows}()},
\code{\link{removecolumns}()},
\code{\link{removeisoinfo}()},
\code{\link{removerows}()},
\code{\link{renamecolumns}()},
\code{\link{rightjoin}()},
\code{\link{runcolfun}()},
\code{\link{selectuniques}()},
\code{\link{selectwherein}()},
\code{\link{selectwhere}()},
\code{\link{setcolclass}()},
\code{\link{setcolnames}()},
\code{\link{setcoltofront}()},
\code{\link{setcolumns}()},
\code{\link{unlistcolumns}()},
\code{\link{updatetablevalues}()}
}
\concept{data.table functions}
