% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{mergeappend}
\alias{mergeappend}
\title{Merge two data tables with preference}
\usage{
mergeappend(x, y, by = NULL, by.x = NULL, by.y = NULL, ...)
}
\description{
Merge two data tables in which the second data.table appends values in the first one, meaning
that \code{NA} values in the first data.table are updated with the values of the second data.table.
Missing values (\code{NA}) in the second data.table will not remove values from the first data.table.
Uses the same variables and functionality as the default data.table \code{\link[data.table]{merge}}
function, but does post-processing on the merged table to overwrite values.
}
