% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{mergeoverwrite}
\alias{mergeoverwrite}
\title{Merge two data tables with overwrite}
\usage{
mergeoverwrite(x, y, by = NULL, by.x = NULL, by.y = NULL, ...)
}
\description{
Merge two data tables in which the second data.table overwrites the values in the first one.
Missing values (\code{NA}) in the second data.table will not remove values from the first data.table.
Uses the same variables and functionality as the default data.table \code{\link[data.table]{merge}}
function, but does post-processing on the merged table to overwrite values.
}
