% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedStrings.R
\name{paste0collapse}
\alias{paste0collapse}
\title{String collapse}
\usage{
paste0collapse(collapsestring, ...)
}
\arguments{
\item{collapsestring}{The string to connect all items.}
}
\description{
Collapse a string with a provided collapse string. Different from \code{\link[base]{paste}}
in that the collapse string character is first.
}
\seealso{
Other string functions: 
\code{\link{str_pad_month}()}
}
\concept{string functions}
