% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedPlotting.R
\name{plot_makecolors}
\alias{plot_makecolors}
\title{Make colors}
\usage{
plot_makecolors(plotdata, plotsettings, whichaxis)
}
\arguments{
\item{plotdata}{The data to use in the plot.}

\item{plotsettings}{The settings to use in the plot.}

\item{whichaxis}{The integer value that indicates the axis used.}
}
\value{
The color settings for each data point (list with a named vector \code{color}).
}
\description{
Make the color settings for each data point in a graph.
}
\seealso{
Other plot functions: 
\code{\link{getdefaulthovertext}()},
\code{\link{getdefaultlabeltext}()},
\code{\link{getdefaultplotsettings}()},
\code{\link{makeplotsettings}()},
\code{\link{plot_makemarkers}()},
\code{\link{plotcandlestick}()},
\code{\link{plotdensityfunctions}()},
\code{\link{plotecdffunctions}()},
\code{\link{plothistogram}()},
\code{\link{plotpiechart}()},
\code{\link{plotscatterwithlink}()},
\code{\link{plotscatter}()},
\code{\link{sampledashboard}()},
\code{\link{updateplotsettings}()}
}
\concept{plot functions}
