% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedPlotting.R
\name{plotdensityfunctions}
\alias{plotdensityfunctions}
\title{Plot density functions}
\usage{
plotdensityfunctions(
  datatable,
  columns = NULL,
  returnplotly = FALSE,
  withlegend = TRUE
)
}
\arguments{
\item{datatable}{The data.table with the data to plot density functions from.}

\item{columns}{The columns to plot density functions for. Default=\code{NULL}, which
plots the density functions for all columns.}
}
\value{
Nothing, just plots the density functions.
}
\description{
Plot density functions based on data from a data.table. Simplified plotting function,
without the use of any plotsettings from \code{\link{getdefaultplotsettings}}.
Uses the \code{\link[stats]{density}} function.
}
\seealso{
Other plot functions: 
\code{\link{getdefaulthovertext}()},
\code{\link{getdefaultlabeltext}()},
\code{\link{getdefaultplotsettings}()},
\code{\link{makeplotsettings}()},
\code{\link{plot_makecolors}()},
\code{\link{plot_makemarkers}()},
\code{\link{plotcandlestick}()},
\code{\link{plotecdffunctions}()},
\code{\link{plothistogram}()},
\code{\link{plotpiechart}()},
\code{\link{plotscatterwithlink}()},
\code{\link{plotscatter}()},
\code{\link{sampledashboard}()},
\code{\link{updateplotsettings}()}
}
\concept{plot functions}
