% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedPlotting.R
\name{plotecdffunctions}
\alias{plotecdffunctions}
\title{Plot ecdf functions}
\usage{
plotecdffunctions(
  datatable,
  columns = NULL,
  quantilerange = c(0, 1),
  inpercentage = FALSE,
  returnplotly = FALSE,
  withlegend = TRUE
)
}
\arguments{
\item{datatable}{The data.table with the data to plot ecdf functions from.}

\item{columns}{The columns to plot ecdf functions for. Default=\code{NULL}, which
plots the density functions for all columns.}

\item{quantilerange}{The range to plot, between 0 and 1, at least two values.
Default=\code{c(0,1)}, which plots the full range.}

\item{inpercentage}{Set to \code{TRUE} to display percentages between 0 and 100.
Default=\code{FALSE}, which displays factors between 0 and 1.}
}
\value{
Nothing, just plots the ecdf functions.
}
\description{
Plot empirical cumulative distribution functions based on data from a data.table.
Uses the \code{\link[stats]{ecdf}} function. Simplified plotting function,
without the use of any plotsettings from \code{\link{getdefaultplotsettings}}.
}
\seealso{
Other plot functions: 
\code{\link{getdefaulthovertext}()},
\code{\link{getdefaultlabeltext}()},
\code{\link{getdefaultplotsettings}()},
\code{\link{makeplotsettings}()},
\code{\link{plot_makecolors}()},
\code{\link{plot_makemarkers}()},
\code{\link{plotcandlestick}()},
\code{\link{plotdensityfunctions}()},
\code{\link{plothistogram}()},
\code{\link{plotpiechart}()},
\code{\link{plotscatterwithlink}()},
\code{\link{plotscatter}()},
\code{\link{sampledashboard}()},
\code{\link{updateplotsettings}()}
}
\concept{plot functions}
