% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedPlotting.R
\name{plotscatterwithlink}
\alias{plotscatterwithlink}
\title{Plot a scatterplot}
\usage{
plotscatterwithlink(
  datatable,
  xaxiscolumn = "x",
  yaxiscolumn = "y",
  namecolumn = NULL,
  linkxaxiscolumn = NULL,
  linkyaxiscolumn = NULL,
  linknamecolumn = NULL,
  plotsettings = NULL,
  returnplotly = FALSE,
  plotlysource = "A"
)
}
\arguments{
\item{datatable}{The datatable in which the data is stored to plot.}

\item{xaxiscolumn}{String with column name to use for the x axis (default=\code{"x"}).}

\item{yaxiscolumn}{String with column name to use for the y axis (default=\code{"y"}).}

\item{namecolumn}{String with column name to use to display name information, meaning that each data point has
a name as specified in this column. Default=\code{NULL}, which does not use point names.}

\item{linkxaxiscolumn}{String with column name to use for the x axis, drawing linkage lines from the representing
\code{xaxiscolumn} and \code{yaxiscolumn}. Default=\code{NULL}, which draws no linkage lines between related points.}

\item{linkyaxiscolumn}{String with column name to use for the y axis, drawing linkage lines from the representing
\code{xaxiscolumn} and \code{yaxiscolumn}. Default=\code{NULL}, which draws no linkage lines between related points.}

\item{linknamecolumn}{String with column name to use to display name information for linkage points. Default=\code{NULL},
which uses the settings as specified in \code{namecolumn}.}

\item{plotsettings}{The settings to use in the plot. Default=\code{NULL}, which loads default plotsettings
from \code{\link{getdefaultplotsettings}}.}

\item{returnplotly}{Set to \code{TRUE} to return the plotly variable to be used or modified in another setting (such as in
a dashboard). Default=\code{FALSE}, which displays the plot.}

\item{plotlysource}{Source attribute as defined in \code{\link[plotly]{plot_ly}}). Default=\code{"A"}.}
}
\value{
The plotly variable or an invisible \code{NULL}.
}
\description{
Plots a scatterplot with columns from a data.table.
}
\seealso{
Other plot functions: 
\code{\link{getdefaulthovertext}()},
\code{\link{getdefaultlabeltext}()},
\code{\link{getdefaultplotsettings}()},
\code{\link{makeplotsettings}()},
\code{\link{plot_makecolors}()},
\code{\link{plot_makemarkers}()},
\code{\link{plotcandlestick}()},
\code{\link{plotdensityfunctions}()},
\code{\link{plotecdffunctions}()},
\code{\link{plothistogram}()},
\code{\link{plotpiechart}()},
\code{\link{plotscatter}()},
\code{\link{sampledashboard}()},
\code{\link{updateplotsettings}()}
}
\concept{plot functions}
