% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedMath.R
\name{rad2deg}
\alias{rad2deg}
\title{Radians to degrees}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{Radians.}
}
\value{
Degrees.
}
\description{
Radians to degrees
}
\seealso{
Other math functions: 
\code{\link{deg2rad}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()}
}
\concept{math functions}
