% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedConversion.R
\name{recasttimezone}
\alias{recasttimezone}
\title{Recast timezone (quick)}
\usage{
recasttimezone(x, tz)
}
\arguments{
\item{x}{The POSIXct time to change}

\item{tz}{The timezone to transform to}
}
\value{
The POSIXct time in the new timezone
}
\description{
Transform a timezone to another but keep the time the same (no calculation).
Uses a numeric difference calculation based on the timezone time difference
of the first item instead of character transformations, so that transformation
is faster. This might not always be the right way when dealing with timezone
values that have a change in daylight saving time.
}
