% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedData.R
\name{removeoffset}
\alias{removeoffset}
\title{Remove offset
Remove offset value from time series data for a period that is considered non-moving.}
\usage{
removeoffset(dataset, statictime = 60, timecolumn = NULL, datacolumn = NULL)
}
\arguments{
\item{dataset}{The data.table containing the data from which offset needs to be removed.}

\item{statictime}{The amount of seconds of the beginning of the dataset to take as static non-moving. Default=\code{60}.}

\item{timecolumn}{The name of the column that contains the time information in POSIXct.
Default=\code{NULL}, which takes the first column.}

\item{datacolumn}{The name of the column(s) that contain the data (numeric). Default=\code{NULL}, which takes all columns
that are not the time column.}
}
\value{
The sensor data with the offset removed, or \code{NULL} on failure.
}
\description{
Remove offset
Remove offset value from time series data for a period that is considered non-moving.
}
\seealso{
Other data functions: 
\code{\link{createbins}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{timeseriesstats}()}

Other time series functions: 
\code{\link{createbins}()},
\code{\link{timeseriesstats}()}
}
\concept{data functions}
\concept{time series functions}
