% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{removerows}
\alias{removerows}
\title{Remove rows by reference}
\usage{
removerows(data, remrows)
}
\arguments{
\item{data}{The data.table to remove the rows from.}

\item{remrows}{Row indices (for example as returned by which=TRUE).}
}
\value{
Returns the modified data.table. It is modified by reference if not all
rows are removed, but returns a 'copy' if all rows are removed.
}
\description{
Remove specific rows from a data.table (by providing indices). Note that removing
a row by reference when the table only consists of one row does not work.
}
\seealso{
Other data.table functions: 
\code{\link{addcolumn}()},
\code{\link{addisoinfo}()},
\code{\link{copytoclipboard}()},
\code{\link{dojoin}()},
\code{\link{extractcategories}()},
\code{\link{fulljoin}()},
\code{\link{getcolumnclasses}()},
\code{\link{innerjoin}()},
\code{\link{leftjoin}()},
\code{\link{loadlargetable}()},
\code{\link{loadsastable}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{makeintegercolumns}()},
\code{\link{makenumericcolumns}()},
\code{\link{mapcolumnclasses}()},
\code{\link{prefixcolnames}()},
\code{\link{removeNAfromtable}()},
\code{\link{removeallNAcolumns}()},
\code{\link{removeallNArows}()},
\code{\link{removecolumns}()},
\code{\link{removeisoinfo}()},
\code{\link{renamecolumns}()},
\code{\link{rightjoin}()},
\code{\link{runcolfun}()},
\code{\link{selectuniques}()},
\code{\link{selectwherein}()},
\code{\link{selectwhere}()},
\code{\link{setcolclass}()},
\code{\link{setcolnames}()},
\code{\link{setcoltofront}()},
\code{\link{setcolumns}()},
\code{\link{unlistcolumns}()},
\code{\link{updatetablevalues}()}
}
\concept{data.table functions}
