% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedVector.R
\name{replaceelements}
\alias{replaceelements}
\title{Replace elements in a vector}
\usage{
replaceelements(input, condition, values)
}
\arguments{
\item{input}{The input vector from which elements need to be replaced.}

\item{condition}{A condition or TRUE/FALSE vector that specifies which
elements need to be replaced (need to be of equal length as input).}

\item{values}{A vector (need to be of equal length as input) for which
values need to be taken where the condition is TRUE.}
}
\value{
New vector in which the elements of input are replaced with elements
from values for which condition is TRUE.
}
\description{
Replace elements in a vector by values from another (equal length) vector based on a vector condition.
}
\seealso{
Other vector functions: 
\code{\link{chunk}()},
\code{\link{dorollapply}()},
\code{\link{extendvector}()},
\code{\link{findconstantchunks}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()},
\code{\link{shift}()}
}
\concept{vector functions}
