% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedVector.R
\name{replaceNAelements}
\alias{replaceNAelements}
\title{Replace NA in a vector}
\usage{
replaceNAelements(input, value = 0)
}
\arguments{
\item{input}{The input vector from which \code{NA} elements need to be replaced.}

\item{value}{The value to use instead of \code{NA} Default=\code{0}.}
}
\value{
New vector in which \code{NA} is replaced by the value.
}
\description{
Replace all \code{NA} values in a vector by a single value.
}
\seealso{
Other vector functions: 
\code{\link{chunk}()},
\code{\link{dorollapply}()},
\code{\link{extendvector}()},
\code{\link{findconstantchunks}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceelements}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()},
\code{\link{shift}()}
}
\concept{vector functions}
