% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rijksdriehoek.R
\name{reproject_geojson}
\alias{reproject_geojson}
\title{Reproject a geoJSON file to an other coordinate system}
\usage{
reproject_geojson(
  filein,
  fileout,
  projection,
  inencoding = getOption("encoding")
)
}
\arguments{
\item{filein}{the input file or connection.}

\item{fileout}{the output file or connection.}

\item{projection}{a function that projects coordinates from the original
coordinate system to the new system. The function has to accept numeric
vectors of length two as its input and return a vector of length two.}

\item{inencoding}{when filein is a character string this encoding is used
when reading from the file.}
}
\value{
A list containing the converted geoJSON. This output is usually not needed.
}
\description{
Reproject a geoJSON file to an other coordinate system
}
