% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedMath.R
\name{runavg}
\alias{runavg}
\title{Running average}
\usage{
runavg(vec, width)
}
\arguments{
\item{vec}{The vector to calculate the running average for.}

\item{width}{Width of the window.}
}
\value{
Vector of the same length as \code{vec} with the running average.
}
\description{
Calculate running average.
}
\seealso{
Other math functions: 
\code{\link{deg2rad}()},
\code{\link{rad2deg}()},
\code{\link{runavgweighed}()},
\code{\link{runsd}()}

Other vector functions: 
\code{\link{dorollapply}()},
\code{\link{extendvector}()},
\code{\link{findconstantchunks}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{replaceelements}()},
\code{\link{runavgweighed}()},
\code{\link{runsd}()},
\code{\link{shift}()}
}
\concept{math functions}
\concept{vector functions}
