% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedMath.R
\name{runavgweighed}
\alias{runavgweighed}
\title{Running average weigthed}
\usage{
runavgweighed(vec, weighvals)
}
\arguments{
\item{vec}{The vector to calculate the running average for.}

\item{weighvals}{Vector containing the weighing values (length of vector indicates the amount of previous points to use for calculation).}
}
\value{
Vector of the same length as \code{vec} with the running average.
}
\description{
Calculate running average with weighing factors.
}
\seealso{
Other math functions: 
\code{\link{deg2rad}()},
\code{\link{rad2deg}()},
\code{\link{runavg}()},
\code{\link{runsd}()}

Other vector functions: 
\code{\link{chunk}()},
\code{\link{dorollapply}()},
\code{\link{extendvector}()},
\code{\link{findconstantchunks}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{replaceelements}()},
\code{\link{runavg}()},
\code{\link{runsd}()},
\code{\link{shift}()}
}
\concept{math functions}
\concept{vector functions}
