% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedInteractive.R
\name{sampledashboard}
\alias{sampledashboard}
\title{Interactive dashboard}
\usage{
sampledashboard(
  alldata,
  plotsettings = NULL,
  categorycolumnname = "CATEGORY",
  stoponlastaction = TRUE
)
}
\arguments{
\item{alldata}{The data.table to interactively display.}

\item{plotsettings}{The settings for the plot. Has to be a modification of the returned list from \code{\link{getdefaultplotsettings}}
default=\code{(NULL}, which plots according to the default settings as returned by \code{\link{getdefaultplotsettings}}).}

\item{categorycolumnname}{String with column name to use for the different categories axis (default=\code{"CATEGORY"}).}

\item{stoponlastaction}{Set to \code{FALSE} to keep the interactive plot running when the last item in the dropdown item has been handled.
Default=\code{TRUE}, which closes the plot after the last item has been handled.}
}
\description{
Display data from a data.table in an interactive graph. This function displays a sample dashboard
with many buttons lacking exact functionality. It's intention is for the code to be
modified to fit a dashboard designed for a specific application.
}
\seealso{
Other plot functions: 
\code{\link{getdefaulthovertext}()},
\code{\link{getdefaultlabeltext}()},
\code{\link{getdefaultplotsettings}()},
\code{\link{makeplotsettings}()},
\code{\link{plot_makecolors}()},
\code{\link{plot_makemarkers}()},
\code{\link{plotcandlestick}()},
\code{\link{plotdensityfunctions}()},
\code{\link{plotecdffunctions}()},
\code{\link{plothistogram}()},
\code{\link{plotpiechart}()},
\code{\link{plotscatterwithlink}()},
\code{\link{plotscatter}()},
\code{\link{updateplotsettings}()}
}
\concept{plot functions}
