% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDatatable.R
\name{selectwherein}
\alias{selectwherein}
\title{Extract from a dataset}
\usage{
selectwherein(data, datacolumn, datavalues, selectcols = NULL, uniques = FALSE)
}
\arguments{
\item{data}{The data.table in which all approaches are listed.}

\item{datacolumn}{String containing the column which must match the values.}

\item{datavalues}{The data values to search for.}

\item{selectcols}{Vector containing column name strings to export (default=\code{NULL}, which returns all columns).}

\item{uniques}{Set to \code{TRUE} if only unique rows are to be returned (default=\code{FALSE}, which returns all matching rows).}
}
\value{
Data.table containing all rows for which the column matches the values.
}
\description{
Extract specified columns from the dataset where 1 specific column contains specific values.
This follows the SQL principle "SELECT selectcols FROM data WHERE datacolumn IN datavalues".
}
\examples{
selectwherein(data, "SIGNAALREMCURVE",c("ATTENTIE","ALARM"),c("DATUM","TREINNR"),TRUE) #returns all unique dates and train numbers where signal is attention or alarm
}
\seealso{
Other data.table functions: 
\code{\link{addcolumn}()},
\code{\link{addisoinfo}()},
\code{\link{copytoclipboard}()},
\code{\link{dojoin}()},
\code{\link{extractcategories}()},
\code{\link{fulljoin}()},
\code{\link{getcolumnclasses}()},
\code{\link{innerjoin}()},
\code{\link{leftjoin}()},
\code{\link{loadlargetable}()},
\code{\link{loadsastable}()},
\code{\link{loadspecifictable}()},
\code{\link{loadtable}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{makeintegercolumns}()},
\code{\link{makenumericcolumns}()},
\code{\link{mapcolumnclasses}()},
\code{\link{prefixcolnames}()},
\code{\link{removeNAfromtable}()},
\code{\link{removeallNAcolumns}()},
\code{\link{removeallNArows}()},
\code{\link{removecolumns}()},
\code{\link{removeisoinfo}()},
\code{\link{removerows}()},
\code{\link{renamecolumns}()},
\code{\link{rightjoin}()},
\code{\link{runcolfun}()},
\code{\link{selectuniques}()},
\code{\link{selectwhere}()},
\code{\link{setcolclass}()},
\code{\link{setcolnames}()},
\code{\link{setcoltofront}()},
\code{\link{setcolumns}()},
\code{\link{unlistcolumns}()},
\code{\link{updatetablevalues}()}
}
\concept{data.table functions}
