% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedDateTime.R
\name{seqdays}
\alias{seqdays}
\title{Sequence of days}
\usage{
seqdays(startdate = NULL, enddate = NULL, stepdate = 1, timezone = "")
}
\arguments{
\item{startdate}{The start date. Default=\code{NULL}, which assumes today.}

\item{enddate}{The end date. Default=\code{NULL}, which assumes today.}

\item{stepdate}{The sequence increment (default=\code{1}).}

\item{timezone}{String determining the timezone (from \code{as.POSIXct}). Default=\code{""},
which uses the local time zone.}
}
\value{
The sequence in POSIXct format.
}
\description{
Get a sequence of days.
}
\seealso{
Other date and time functions: 
\code{\link{addisoinfo}()},
\code{\link{makeconsistentdateformat}()},
\code{\link{makeconsistentdatetimeformat}()},
\code{\link{makedatecolumns}()},
\code{\link{makedatetimecolumns}()},
\code{\link{removeisoinfo}()}
}
\concept{date and time functions}
