% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedVector.R
\name{shift}
\alias{shift}
\title{Shift vector values}
\usage{
shift(x, n, invert = FALSE, default = NA)
}
\arguments{
\item{x}{Vector for which to shift values.}

\item{n}{Number of places to be shifted.

Positive numbers will shift to the right while negative numbers will shift to the left.
This direction can be inverted by the invert parameter.}

\item{invert}{Whether shift direction should be inverted.}

\item{default}{The value that should be inserted by default (at the empty spots).}
}
\value{
The shifted vector.
}
\description{
Shift vector values to the left or right.
}
\seealso{
Other vector functions: 
\code{\link{dorollapply}()},
\code{\link{extendvector}()},
\code{\link{findconstantchunks}()},
\code{\link{findconstant}()},
\code{\link{leaveonlychunks}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{makelength}()},
\code{\link{replaceNAelements}()},
\code{\link{replaceelements}()},
\code{\link{runavgweighed}()},
\code{\link{runavg}()},
\code{\link{runsd}()}
}
\concept{vector functions}
