% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedStrings.R
\name{str_pad_month}
\alias{str_pad_month}
\title{Pad month with 0}
\usage{
str_pad_month(themonth)
}
\arguments{
\item{themonth}{Month value or string to pad}
}
\value{
String with month value padded (always zero padded left, length 2)
}
\description{
Simple function to pad zeros to a month
}
\seealso{
Other string functions: 
\code{\link{paste0collapse}()}
}
\concept{string functions}
