% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedLists.R
\name{stripnames}
\alias{stripnames}
\title{Strip string of named elements}
\usage{
stripnames(thelist, thenames)
}
\arguments{
\item{thelist}{List to remove names from.}

\item{thenames}{Vector of strings containing the names of the elements to remove from the list.}
}
\value{
List with all the named elements removed.
}
\description{
Remove all specific named elements from a list. Executes recursively so
that it also works within lists in lists.
}
\seealso{
Other list functions: 
\code{\link{forcelisttype}()},
\code{\link{leavenames}()},
\code{\link{listdepth}()}
}
\concept{list functions}
