% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedData.R
\name{timeseriesstats}
\alias{timeseriesstats}
\title{Get some time series statistics}
\usage{
timeseriesstats(thetimes, thedata = NULL)
}
\arguments{
\item{thetimes}{Data.table with the first column being the DateTime and the second column being the values,
or, a vector containing the DateTime only.}

\item{thedata}{Vector containing the values correlated to the DateTime. Must be of equal length to \code{thetimes}.
Default=\code{NULL}, which only uses the first parameter.}
}
\value{
A named list with some time series statistics or \code{NULL} on failure. The "grouping" variables show
how many points would be within a specific time period, and how many bins would occur for the full time period.
This can be useful for calculating optimal statistics for data lasting a long time.
}
\description{
Get some time series statistics
}
\seealso{
createbins

Other data functions: 
\code{\link{createbins}()},
\code{\link{localmaxima}()},
\code{\link{localminimamaxima}()},
\code{\link{localminima}()},
\code{\link{removeoffset}()}

Other time series functions: 
\code{\link{createbins}()},
\code{\link{removeoffset}()}
}
\concept{data functions}
\concept{time series functions}
