% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuicedPlotting.R
\name{updateplotsettings}
\alias{updateplotsettings}
\title{Update plotsettings}
\usage{
updateplotsettings(plotsettings, datacolnames)
}
\arguments{
\item{plotsettings}{The plotsettings to check.}

\item{datacolnames}{The column names present in the data.}
}
\value{
Updated plotsettings with axes removed that are not present in the data. If none of the axes
are present (either x-axis or y-axes), the function will return \code{NULL}.
}
\description{
Update the plotsettings based on the column names in the data. It checks if the provided settings
are possible to display. This function is always called from the plot functions.
}
\seealso{
Other plot functions: 
\code{\link{getdefaulthovertext}()},
\code{\link{getdefaultlabeltext}()},
\code{\link{getdefaultplotsettings}()},
\code{\link{makeplotsettings}()},
\code{\link{plot_makecolors}()},
\code{\link{plot_makemarkers}()},
\code{\link{plotcandlestick}()},
\code{\link{plotdensityfunctions}()},
\code{\link{plotecdffunctions}()},
\code{\link{plothistogram}()},
\code{\link{plotpiechart}()},
\code{\link{plotscatterwithlink}()},
\code{\link{plotscatter}()},
\code{\link{sampledashboard}()}
}
\concept{plot functions}
