% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rijksdriehoek.R
\name{wgs84_to_rd}
\alias{wgs84_to_rd}
\title{Translate coordinates in WGS84 to rijksdriehoek (Dutch National Grid)}
\usage{
wgs84_to_rd(lambda, phi)
}
\arguments{
\item{lambda}{numeric vector with longitutes (in decimal degrees), or
without phi a vector of length 2 containing both phi and lambda.}

\item{phi}{numeric vector with latitudes (in decimal degrees).}
}
\value{
A list with x and y coordinates in Rijksdriehoek system coordinates, of if
just lambda was given a numeric vector of length 2 with both coordinates.
}
\description{
Translate coordinates in WGS84 to rijksdriehoek (Dutch National Grid)
}
